ia_prompt_music/requirements.txt
python-dotenv>=1.0.0
librosa>=0.10.1
pydub>=0.25.1
numpy>=1.24.4
soundfile>=0.12.1
google-generativeai>=0.3.2 